+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `laravel-doctrine` will be documented in this file
 
+
+## 1.0.1 - 2021-05-17
+### What's New
+Add namespaces to generated entities, with config parameter
+### Improvements
+Reformat Traits
+
 ## 1.0.0 - 2021-04-21
 
 - initial release